+++ ormar/models/descriptors/descriptors.py
@@ -4,10 +4,7 @@ from typing import Any, TYPE_CHECKING, Tuple, Type, Union, cast
 
 import ormar
 
-try:
-    import orjson as json
-except ImportError:  # pragma: no cover
-    import json  # type: ignore
+from ormar.fields.parsers import encode_json
 
 if TYPE_CHECKING:  # pragma: no cover
     from ormar import Model
@@ -43,9 +40,7 @@ class JsonDescriptor:
         return value
 
     def __set__(self, instance: "Model", value: Any) -> None:
-        if not isinstance(value, str):
-            value = json.dumps(value)
-        value = value.decode("utf-8") if isinstance(value, bytes) else value
+        value = encode_json(value)
         instance._internal_set(self.name, value)
         instance.set_save_status(False)
 
@@ -62,7 +57,11 @@ class BytesDescriptor:
     def __get__(self, instance: "Model", owner: Type["Model"]) -> Any:
         value = instance.__dict__.get(self.name, None)
         field = instance.Meta.model_fields[self.name]
-        if field.represent_as_base64_str and not isinstance(value, str):
+        if (
+            value is not None
+            and field.represent_as_base64_str
+            and not isinstance(value, str)
+        ):
             value = base64.b64encode(value).decode()
         return value