+++ ormar/models/mixins/save_mixin.py
@@ -1,3 +1,4 @@
+import base64
 from typing import (
     Any,
     Callable,
@@ -13,6 +14,7 @@ from typing import (
 import pydantic
 
 import ormar  # noqa: I100, I202
+from ormar.fields.parsers import encode_json
 from ormar.models.mixins import AliasMixin
 from ormar.models.mixins.relation_mixin import RelationMixin
 
@@ -30,6 +32,8 @@ class SavePrepareMixin(RelationMixin, AliasMixin):
         _skip_ellipsis: Callable
         pk_names_list: List[str]
         pk_aliases_list: List[str]
+        _json_fields: Set[str]
+        _bytes_fields: Set[str]
         __fields__: Dict[str, pydantic.fields.ModelField]
 
     @classmethod
@@ -52,6 +56,25 @@ class SavePrepareMixin(RelationMixin, AliasMixin):
         new_kwargs = cls._remove_not_ormar_fields(new_kwargs)
         new_kwargs = cls.substitute_models_with_pks(new_kwargs)
         new_kwargs = cls.populate_default_values(new_kwargs)
+        new_kwargs = cls.reconvert_str_to_bytes(new_kwargs)
+        new_kwargs = cls.dump_all_json_fields_to_str(new_kwargs)
+        new_kwargs = cls.translate_columns_to_aliases(new_kwargs)
+        return new_kwargs
+
+    @classmethod
+    def prepare_model_to_update(cls, new_kwargs: dict) -> dict:
+        """
+        Combines all preparation methods before updating.
+        :param new_kwargs: dictionary of model that is about to be saved
+        :type new_kwargs: Dict[str, str]
+        :return: dictionary of model that is about to be updated
+        :rtype: Dict[str, str]
+        """
+        new_kwargs = cls.populate_compound_pk(new_kwargs)
+        new_kwargs = cls.parse_non_db_fields(new_kwargs)
+        new_kwargs = cls.substitute_models_with_pks(new_kwargs)
+        new_kwargs = cls.reconvert_str_to_bytes(new_kwargs)
+        new_kwargs = cls.dump_all_json_fields_to_str(new_kwargs)
         new_kwargs = cls.translate_columns_to_aliases(new_kwargs)
         return new_kwargs
 
@@ -132,7 +155,10 @@ class SavePrepareMixin(RelationMixin, AliasMixin):
                 parsers = {"string": lambda x: str(x), "hex": lambda x: "%.32x" % x.int}
                 uuid_format = field.column_type.uuid_format
                 parser: Callable[..., Any] = parsers.get(uuid_format, lambda x: x)
-                model_dict[name] = parser(model_dict[name])
+                if isinstance(model_dict[name], dict):
+                    model_dict[name] = field.to.parse_non_db_fields(model_dict[name])
+                else:
+                    model_dict[name] = parser(model_dict[name])
         return model_dict
 
     @classmethod
@@ -167,6 +193,47 @@ class SavePrepareMixin(RelationMixin, AliasMixin):
         return model_dict
 
     @classmethod
+    def reconvert_str_to_bytes(cls, model_dict: Dict) -> Dict:
+        """
+        Receives dictionary of model that is about to be saved and changes
+        all bytes fields that are represented as strings back into bytes.
+
+        :param model_dict: dictionary of model that is about to be saved
+        :type model_dict: Dict
+        :return: dictionary of model that is about to be saved
+        :rtype: Dict
+        """
+        bytes_base64_fields = {
+            name
+            for name, field in cls.Meta.model_fields.items()
+            if field.represent_as_base64_str
+        }
+        for key, value in model_dict.items():
+            if key in cls._bytes_fields and isinstance(value, str):
+                model_dict[key] = (
+                    value.encode("utf-8")
+                    if key not in bytes_base64_fields
+                    else base64.b64decode(value)
+                )
+        return model_dict
+
+    @classmethod
+    def dump_all_json_fields_to_str(cls, model_dict: Dict) -> Dict:
+        """
+        Receives dictionary of model that is about to be saved and changes
+        all json fields into strings
+
+        :param model_dict: dictionary of model that is about to be saved
+        :type model_dict: Dict
+        :return: dictionary of model that is about to be saved
+        :rtype: Dict
+        """
+        for key, value in model_dict.items():
+            if key in cls._json_fields:
+                model_dict[key] = encode_json(value)
+        return model_dict
+
+    @classmethod
     def populate_default_values(cls, new_kwargs: Dict) -> Dict:
         """
         Receives dictionary of model that is about to be saved and populates the default
@@ -186,7 +253,10 @@ class SavePrepareMixin(RelationMixin, AliasMixin):
             ):
                 new_kwargs[field_name] = field.get_default()
             # clear fields with server_default set as None
-            if field.server_default is not None and not new_kwargs.get(field_name):
+            if (
+                field.server_default is not None
+                and new_kwargs.get(field_name, None) is None
+            ):
                 new_kwargs.pop(field_name, None)
         return new_kwargs