+++ ormar/models/newbasemodel.py
@@ -79,7 +79,6 @@ class NewBaseModel(pydantic.BaseModel, ModelTableProxy, metaclass=ModelMetaclass
         pk: Any
         __model_fields__: Dict[str, BaseField]
         __table__: sqlalchemy.Table
-        __fields__: Dict[str, pydantic.fields.ModelField]
         __pydantic_model__: Type[BaseModel]
         __pkname__: str
         __tablename__: str
@@ -198,6 +197,29 @@ class NewBaseModel(pydantic.BaseModel, ModelTableProxy, metaclass=ModelMetaclass
         :rtype: Any
         """
         return super().__getattribute__(item)
+
+    def __getstate__(self) -> Dict[Any, Any]:
+        state = super().__getstate__()
+        self_dict = self.dict()
+        state["__dict__"].update(**self_dict)
+        return state
+
+    def __setstate__(self, state: Dict[Any, Any]) -> None:
+        relations = {
+            k: v
+            for k, v in state["__dict__"].items()
+            if k in self.extract_related_names()
+        }
+        basic_state = {
+            k: v
+            for k, v in state["__dict__"].items()
+            if k not in self.extract_related_names()
+        }
+        state["__dict__"] = basic_state
+        super().__setstate__(state)
+        self._initialize_internal_attributes()
+        for name, value in relations.items():
+            setattr(self, name, value)
 
     def _internal_set(self, name: str, value: Any) -> None:
         """
@@ -892,7 +914,7 @@ class NewBaseModel(pydantic.BaseModel, ModelTableProxy, metaclass=ModelMetaclass
         if column_name not in self._bytes_fields:
             return value
         field = self.Meta.model_fields[column_name]
-        if not isinstance(value, bytes):
+        if not isinstance(value, bytes) and value is not None:
             if field.represent_as_base64_str:
                 value = base64.b64decode(value)
             else:
@@ -913,7 +935,11 @@ class NewBaseModel(pydantic.BaseModel, ModelTableProxy, metaclass=ModelMetaclass
         if column_name not in self._bytes_fields:
             return value
         field = self.Meta.model_fields[column_name]
-        if not isinstance(value, str) and field.represent_as_base64_str:
+        if (
+            value is not None
+            and not isinstance(value, str)
+            and field.represent_as_base64_str
+        ):
             return base64.b64encode(value).decode()
         return value