+++ ormar/queryset/actions/filter_action.py
@@ -1,8 +1,6 @@
-import datetime
 from typing import Any, Dict, TYPE_CHECKING, Tuple, Type
 
 import sqlalchemy
-from sqlalchemy import text
 from sqlalchemy.sql.elements import BooleanClauseList
 
 import ormar  # noqa: I100, I202
@@ -133,43 +131,6 @@ class FilterAction(QueryAction):
         aliased_name = relation_field.names.get(column_name)
         return self.target_model.Meta.table.columns[aliased_name]
 
-    def get_text_clause(self) -> sqlalchemy.sql.expression.TextClause:
-        """
-        Escapes characters if it's required.
-        Substitutes values of the models if value is a ormar Model with its pk value.
-        Compiles the clause.
-
-        :return: complied and escaped clause
-        :rtype: sqlalchemy.sql.elements.TextClause
-        """
-        if isinstance(self.filter_value, ormar.Model):
-            self.filter_value = self.filter_value.pk
-
-        op_attr, filter_value = self._get_operator_attribute_and_filter_value()
-
-        if isinstance(filter_value, dict):
-            clause = self._prepare_composite_primary_keys_filter(
-                op_attr=op_attr, filter_value=filter_value
-            )
-        else:
-            clause = getattr(self.column, op_attr)(filter_value)
-            clause = self._compile_clause(
-                clause,
-                modifiers={"escape": "\\" if self.has_escaped_character else None},
-            )
-        return clause
-
-    def _get_operator_attribute_and_filter_value(self) -> Tuple[str, Any]:
-        op_attr = FILTER_OPERATORS[self.operator]
-
-        self._convert_dates_if_required()
-        filter_value = self.filter_value
-
-        if self.operator == "isnull":
-            op_attr = "is_" if self.filter_value else "isnot"
-            filter_value = None
-        return op_attr, filter_value
-
     def _prepare_composite_primary_keys_filter(
         self, op_attr: str, filter_value: Dict
     ) -> BooleanClauseList:
@@ -181,63 +142,43 @@ class FilterAction(QueryAction):
             getattr(self._get_composite_pk_column(key), op_attr)(filter_value[key])
             for key in filter_value.keys()
         ]
-        clauses = [
-            self._compile_clause(
-                clause,
-                modifiers={"escape": "\\" if self.has_escaped_character else None},
-            )
-            for clause in clauses
-        ]
         return sqlalchemy.sql.and_(*clauses)
 
-    def _convert_dates_if_required(self) -> None:
+    def get_text_clause(self) -> sqlalchemy.sql.expression.BinaryExpression:
         """
-        Converts dates, time and datetime to isoformat
-        """
-        if isinstance(
-            self.filter_value, (datetime.date, datetime.time, datetime.datetime)
-        ):
-            self.filter_value = self.filter_value.isoformat()
-
-        if isinstance(self.filter_value, (list, tuple, set)):
-            self.filter_value = [
-                x.isoformat()
-                if isinstance(x, (datetime.date, datetime.time, datetime.datetime))
-                else x
-                for x in self.filter_value
-            ]
-
-    def _compile_clause(
-        self, clause: sqlalchemy.sql.expression.BinaryExpression, modifiers: Dict
-    ) -> sqlalchemy.sql.expression.TextClause:
-        """
-        Compiles the clause to str using appropriate database dialect, replace columns
-        names with aliased names and converts it back to TextClause.
-
-        :param clause: original not compiled clause
-        :type clause: sqlalchemy.sql.elements.BinaryExpression
-        :param modifiers: sqlalchemy modifiers - used only to escape chars here
-        :type modifiers: Dict[str, NoneType]
-        :return: compiled and escaped clause
+        Escapes characters if it's required.
+        Substitutes values of the models if value is a ormar Model with its pk value.
+        Compiles the clause.
+        :return: complied and escaped clause
         :rtype: sqlalchemy.sql.elements.TextClause
         """
-        for modifier, modifier_value in modifiers.items():
-            clause.modifiers[modifier] = modifier_value
+        if isinstance(self.filter_value, ormar.Model):
+            self.filter_value = self.filter_value.pk
 
-        column = clause.left
-        clause_text = str(
-            clause.compile(
-                dialect=self.target_model.Meta.database._backend._dialect,
-                compile_kwargs={"literal_binds": True},
+        op_attr = FILTER_OPERATORS[self.operator]
+
+        if self.operator == "isnull":
+            op_attr = "is_" if self.filter_value else "isnot"
+            filter_value = None
+        else:
+            filter_value = self.filter_value
+
+        if isinstance(filter_value, dict):
+            clause = self._prepare_composite_primary_keys_filter(
+                op_attr=op_attr, filter_value=filter_value
             )
-        )
-        alias = f"{self.table_prefix}_" if self.table_prefix else ""
-        aliased_name = f"{alias}{self.table.name}.{column.name}"
-        clause_text = clause_text.replace(
-            f"{self.table.name}.{column.name}", aliased_name
-        )
-        dialect_name = self.target_model.Meta.database._backend._dialect.name
-        if dialect_name != "sqlite":  # pragma: no cover
-            clause_text = clause_text.replace("%%", "%")  # remove %% in some dialects
-        clause = text(clause_text)
+        else:
+            if self.table_prefix:
+                aliased_table = (
+                    self.source_model.Meta.alias_manager.prefixed_table_name(
+                        self.table_prefix, self.column.table
+                    )
+                )
+                aliased_column = getattr(aliased_table.c, self.column.name)
+            else:
+                aliased_column = self.column
+            clause = getattr(aliased_column, op_attr)(filter_value)
+
+        if self.has_escaped_character:
+            clause.modifiers["escape"] = "\\"
         return clause