+++ ormar/queryset/queryset.py
@@ -18,9 +18,22 @@ import databases
 import sqlalchemy
 from sqlalchemy import bindparam
 
+try:
+    from sqlalchemy.engine import LegacyRow
+except ImportError:  # pragma: no cover
+    if TYPE_CHECKING:
+
+        class LegacyRow(dict):  # type: ignore
+            pass
+
+
 import ormar  # noqa I100
 from ormar import MultipleMatches, NoMatch
-from ormar.exceptions import ModelPersistenceError, QueryDefinitionError
+from ormar.exceptions import (
+    ModelListEmptyError,
+    ModelPersistenceError,
+    QueryDefinitionError,
+)
 from ormar.queryset import FieldAccessor, FilterQuery, SelectAction
 from ormar.queryset.actions.order_action import OrderAction
 from ormar.queryset.clause import FilterGroup, QueryClause
@@ -605,7 +618,9 @@ class QuerySet(Generic[T]):
             model_cls=self.model_cls,  # type: ignore
             exclude_through=exclude_through,
         )
-        column_map = alias_resolver.resolve_columns(columns_names=list(rows[0].keys()))
+        column_map = alias_resolver.resolve_columns(
+            columns_names=list(cast(LegacyRow, rows[0]).keys())
+        )
         result = [
             {column_map.get(k): v for k, v in dict(x).items() if k in column_map}
             for x in rows
@@ -873,13 +888,17 @@ class QuerySet(Generic[T]):
 
         expr = self.build_select_expression(
             limit=1,
-            order_bys=[
-                OrderAction(
-                    order_str=f"{pkname}",
-                    model_cls=self.model_cls,  # type: ignore
-                )
-                for pkname in self.model_cls.pk_names_list  # type: ignore
-            ]
+            order_bys=(
+                [
+                    OrderAction(
+                        order_str=f"{pkname}",
+                        model_cls=self.model_cls,  # type: ignore
+                    )
+                    for pkname in self.model_cls.pk_names_list  # type: ignore
+                ]
+                if not any([x.is_source_model_order for x in self.order_bys])
+                else []
+            )
             + self.order_bys,
         )
         rows = await self.database.fetch_all(expr)
@@ -910,7 +929,7 @@ class QuerySet(Generic[T]):
         except ormar.NoMatch:
             return None
 
-    async def get(self, *args: Any, **kwargs: Any) -> "T":
+    async def get(self, *args: Any, **kwargs: Any) -> "T":  # noqa: CCR001
         """
         Get's the first row from the db meeting the criteria set by kwargs.
 
@@ -932,13 +951,17 @@ class QuerySet(Generic[T]):
         if not self.filter_clauses:
             expr = self.build_select_expression(
                 limit=1,
-                order_bys=[
-                    OrderAction(
-                        order_str=f"-{pkname}",
-                        model_cls=self.model_cls,  # type: ignore
-                    )
-                    for pkname in self.model_cls.pk_names_list  # type: ignore
-                ]
+                order_bys=(
+                    [
+                        OrderAction(
+                            order_str=f"-{pkname}",
+                            model_cls=self.model_cls,  # type: ignore
+                        )
+                        for pkname in self.model_cls.pk_names_list  # type: ignore
+                    ]
+                    if not any([x.is_source_model_order for x in self.order_bys])
+                    else []
+                )
                 + self.order_bys,
             )
         else:
@@ -1032,7 +1055,7 @@ class QuerySet(Generic[T]):
 
     async def bulk_create(self, objects: List["T"]) -> None:
         """
-        Performs a bulk update in one database session to speed up the process.
+        Performs a bulk create in one database session to speed up the process.
 
         Allows you to create multiple objects at once.
 
@@ -1043,17 +1066,15 @@ class QuerySet(Generic[T]):
         :param objects: list of ormar models already initialized and ready to save.
         :type objects: List[Model]
         """
-        ready_objects = []
-        for objt in objects:
-            new_kwargs = objt.dict()
-            new_kwargs = objt.prepare_model_to_save(new_kwargs)
-            ready_objects.append(new_kwargs)
+        ready_objects = [obj.prepare_model_to_save(obj.dict()) for obj in objects]
 
-        expr = self.table.insert()
-        await self.database.execute_many(expr, ready_objects)
+        # don't use execute_many, as in databases it's executed in a loop
+        # instead of using execute_many from drivers
+        expr = self.table.insert().values(ready_objects)
+        await self.database.execute(expr)
 
-        for objt in objects:
-            objt.set_save_status(True)
+        for obj in objects:
+            obj.set_save_status(True)
 
     async def bulk_update(  # noqa:  CCR001
         self, objects: List["T"], columns: List[str] = None
@@ -1061,7 +1082,7 @@ class QuerySet(Generic[T]):
         """
         Performs bulk update in one database session to speed up the process.
 
-        Allows to update multiple instance at once.
+        Allows you to update multiple instance at once.
 
         All `Models` passed need to have primary key column populated.
 
@@ -1075,6 +1096,9 @@ class QuerySet(Generic[T]):
         :param columns: list of columns to update
         :type columns: List[str]
         """
+        if not objects:
+            raise ModelListEmptyError("Bulk update objects are empty!")
+
         ready_objects = []
         # TODO: Adjust for multi pk
         pk_names = self.model.pk_names_list
@@ -1090,23 +1114,21 @@ class QuerySet(Generic[T]):
                 columns.append(pk_name)
 
         columns = [self.model.get_column_alias(k) for k in columns]
-        kwargs = dict()
-        for objt in objects:
-            kwargs = objt.dict()
+        new_kwargs = dict()
+        for obj in objects:
+            new_kwargs = obj.dict()
             for pk_name in pk_names:
-                if pk_name not in kwargs or kwargs.get(pk_name) is None:
+                if pk_name not in new_kwargs or new_kwargs.get(pk_name) is None:
                     raise ModelPersistenceError(
                         "You cannot update unsaved objects. "
                         f"{self.model.__name__} has to have {pk_name} filled."
                     )
-            kwargs = self.model.populate_compound_pk(kwargs)
-            kwargs = self.model.substitute_models_with_pks(kwargs)
-            kwargs = self.model.parse_non_db_fields(kwargs)
-            kwargs = self.model.translate_columns_to_aliases(kwargs)
-            new_kwargs = {"new_" + k: v for k, v in kwargs.items() if k in columns}
-            ready_objects.append(new_kwargs)
+            new_kwargs = obj.prepare_model_to_update(new_kwargs)
+            ready_objects.append(
+                {"new_" + k: v for k, v in new_kwargs.items() if k in columns}
+            )
 
-        used_columns = [self.model.get_column_alias(k) for k in kwargs]
+        used_columns = [self.model.get_column_alias(k) for k in new_kwargs]
         table_columns = [
             c.name for c in self.model_meta.table.c if c.name in used_columns
         ]
@@ -1126,5 +1148,9 @@ class QuerySet(Generic[T]):
         expr = str(expr)
         await self.database.execute_many(expr, ready_objects)
 
-        for objt in objects:
-            objt.set_save_status(True)
+        for obj in objects:
+            obj.set_save_status(True)
+
+        await cast(Type["Model"], self.model_cls).Meta.signals.post_bulk_update.send(
+            sender=self.model_cls, instances=objects  # type: ignore
+        )